% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackageForSNPanalysis.R
\name{checkInstallation}
\alias{checkInstallation}
\title{Check whether each package is installed}
\usage{
checkInstallation(package0)
}
\arguments{
\item{package0}{}
}
\description{
Check whether each package is installed
}
