% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function_in_cluster_analysis.R
\name{chooseStrain}
\alias{chooseStrain}
\title{Choose strains which belong to the same group}
\usage{
chooseStrain(type, strain0 = strain_classification)
}
\arguments{
\item{type}{A string representing the strain group name}

\item{strain0}{A dataframe contain all 1011 yeast strains' classification information}
}
\value{
A vector containing strain name
}
\description{
This function is mainly used to choose the strain based on the strain type defined in 1011 genome sequence project
}
\examples{
strain_type <- "all_strain"
strain_select1 <- chooseStrain(type = strain_type)
}
