% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUMPS.R
\name{clumpsAnalysis}
\alias{clumpsAnalysis}
\title{CLUMPS function used to run the CLUMPS pipeline}
\usage{
clumpsAnalysis(gene0, SNPlist0, gene_annotation0, pdbID0, sstart0, send0)
}
\arguments{
\item{send0}{}
}
\description{
CLUMPS function used to run the CLUMPS pipeline
}
