% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{countMutationProtein}
\alias{countMutationProtein}
\title{Count residues's mutation number}
\usage{
countMutationProtein(gene_name, mutation_annotation, gene_snp0)
}
\arguments{
\item{gene_name}{A string containing the systematic gene name}

\item{mutation_annotation}{A dataframe containing the annotation of each SNP for one gene}

\item{gene_snp0}{A list containing the detailed annotation of one gene}
}
\value{
A vector containing the mutation number of residues in each site
}
\description{
Calculate the number of residues mutation in each site
}
