% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{countMutationProtein0}
\alias{countMutationProtein0}
\title{function to obtain the mutation of amino acids residue in the 3d structure based on sequence blast anlysis
but it can be very dangeous using this method, as the insertion or deletion could lead to many mutation in a seq}
\usage{
countMutationProtein0(geneName, mutated_gene_seq)
}
\arguments{
\item{mutated_gene_seq}{}
}
\description{
function to obtain the mutation of amino acids residue in the 3d structure based on sequence blast anlysis
but it can be very dangeous using this method, as the insertion or deletion could lead to many mutation in a seq
}
