% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{filterMetabolicGene}
\alias{filterMetabolicGene}
\title{this function is used to filter fasta file of genes which belong to metabolic}
\usage{
filterMetabolicGene()
}
\description{
this function is used to filter fasta file of genes which belong to metabolic
}
