% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{filterMutationStrainType}
\alias{filterMutationStrainType}
\title{this function is used to choose the gene based on strain phenotype information}
\usage{
filterMutationStrainType(gene_test, strain_select)
}
\arguments{
\item{strain_select}{}
}
\description{
this function is used to choose the gene based on strain phenotype information
}
