% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{findPPosition}
\alias{findPPosition}
\title{this function is used to find the postion of mutated amino acids based on genomics mutation
version2}
\usage{
findPPosition(mutatedPosition, alted, geneName, gene_snp0)
}
\arguments{
\item{geneName}{}
}
\description{
this function is used to find the postion of mutated amino acids based on genomics mutation
version2
}
