% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{findPPosition0}
\alias{findPPosition0}
\title{the followed two function  were used to estimate the mutation information based on input DNA fasta file
---------------version 1}
\usage{
findPPosition0(alted_seq, geneName)
}
\arguments{
\item{geneName}{}
}
\description{
the followed two function  were used to estimate the mutation information based on input DNA fasta file
---------------version 1
}
