% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getDomainCoordinate}
\alias{getDomainCoordinate}
\title{function to get the coordinate information of domain occured in each gene}
\usage{
getDomainCoordinate(dataframe0)
}
\arguments{
\item{dataframe0}{}
}
\description{
function to get the coordinate information of domain occured in each gene
}
