% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneCoordinate.R
\name{getGeneCoordinate}
\alias{getGeneCoordinate}
\title{the function is used to obtain a list which contains the coordinate information of each gene}
\usage{
getGeneCoordinate(gene_name, genesum = gene_feature_GEM)
}
\arguments{
\item{genesum}{}
}
\description{
the function is used to obtain a list which contains the coordinate information of each gene
}
