% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneName.R
\name{getGeneName}
\alias{getGeneName}
\title{function get the gene name based on the mutation position}
\usage{
getGeneName(chr, mutated_positions, gene_annotation = gene_feature0)
}
\arguments{
\item{gene_annotation}{}
}
\description{
function get the gene name based on the mutation position
}
