% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getGeneNameWithSNP}
\alias{getGeneNameWithSNP}
\title{this function is used to obtain the metabolic gene list which have SNP, in total there are 36 metabolic genes which
don't have the mutation}
\usage{
getGeneNameWithSNP()
}
\description{
this function is used to obtain the metabolic gene list which have SNP, in total there are 36 metabolic genes which
don't have the mutation
}
