% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getHotVertice}
\alias{getHotVertice}
\title{this function is used to get the vertices}
\usage{
getHotVertice(aa_3d, residue0, aa_pro, distance0)
}
\arguments{
\item{distance0}{}
}
\description{
this function is used to get the vertices
}
