% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{getMultipleMatchParameter}
\alias{getMultipleMatchParameter}
\title{get multiple mapping results for one element}
\usage{
getMultipleMatchParameter(description, reaction, ko)
}
\arguments{
\item{description}{vector contains the description for each element from 'reaction'}

\item{reaction}{vector contrains the element been mapped}

\item{ko}{vector contains the element need mapping}
}
\value{
one element, which contains several mapping results seperated by ";"
}
\description{
get multiple mapping results for one element
}
\examples{
w <- 1:6
v <- c('z','a','b','a','b', 'e')
testData <- c('a','b','g')
getMultipleMatchParameter(w,v,testData)

}
