% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getOriginalCoordinateProtein}
\alias{getOriginalCoordinateProtein}
\title{this function is used to get the original residues coordinates for the hotspot analysis based on the protein structure}
\usage{
getOriginalCoordinateProtein(coordinate0, coordinate_mapping0)
}
\arguments{
\item{coordinate_mapping0}{}
}
\description{
this function is used to get the original residues coordinates for the hotspot analysis based on the protein structure
}
