% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getProteinFromCDS}
\alias{getProteinFromCDS}
\title{this function can produce the amino acid seq based on cds sequence}
\usage{
getProteinFromCDS(cds0)
}
\arguments{
\item{cds0}{}
}
\description{
this function can produce the amino acid seq based on cds sequence
}
