% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getPvalue}
\alias{getPvalue}
\title{function to calculate the right-tailed p value}
\usage{
getPvalue(wap_initial, wap_sampling)
}
\arguments{
\item{wap_sampling}{}
}
\description{
function to calculate the right-tailed p value
}
