% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{getSingleMatchParameter}
\alias{getSingleMatchParameter}
\title{get single mapping result for one element}
\usage{
getSingleMatchParameter(description, reaction, ko)
}
\arguments{
\item{description}{vector contains the description for each element from 'reaction'}

\item{reaction}{vector contrains the element been mapped}

\item{ko}{vector contains the element need mapping}
}
\value{
one element
}
\description{
get single mapping result for one element
}
\examples{
w <- 1:6
v <- c('z','a','b','a','b', 'e')
testData <- c('a','b','g')
getSingleMatchParameter(w,v,testData)

}
