% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotSpotAnalysis.R
\name{hotSpotAnalysis}
\alias{hotSpotAnalysis}
\title{This function is used to conduct the hotspot analysis}
\usage{
hotSpotAnalysis(gene0 = ss0, pdbID = NA, SNPlist0 = mutated_gene1,
  gene_annotation0 = gene_feature0, pdb = distance, sstart0 = p1,
  send0 = p2, qstart0 = q1, qend0 = q2, result_dir = outfile0,
  strain_type = NA, input_dir = TRUE)
}
\arguments{
\item{gene0}{A gene systematic name}

\item{pdbID}{A string}

\item{SNPlist0}{A SNP list for the strains from specific phenotype}

\item{gene_annotation0}{The gene annotation summary}

\item{pdb}{The dir store the residue distance matrix, should be a txt file seperated by ',' or the residue distance matrix}

\item{sstart0}{The start residue coordinate for the resdiues in the original protein}

\item{send0}{The end residue coordinate for the residues in the the original protein}

\item{qstart0}{The start residue coordinate for the resdiues in the PDB file}

\item{qend0}{The end residue coordinate for the residues in the PDB file}

\item{result_dir}{The directory to save the hot spot analysis result}

\item{strain_type}{A string to represent the source of SNP}

\item{input_dir}{}
}
\description{
This function is used to conduct the hotspot analysis
}
