% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackageForSNPanalysis.R
\name{loadPackageForSNPanalysis}
\alias{loadPackageForSNPanalysis}
\title{This function is used to load the package used for SNP analysis}
\usage{
loadPackageForSNPanalysis()
}
\description{
This function is used to load the package used for SNP analysis
}
