% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{mappingCoordinateFrom3DtoProtein}
\alias{mappingCoordinateFrom3DtoProtein}
\title{this function is used to establish the mapping in the coordinate from protein structure and from protein sequence}
\usage{
mappingCoordinateFrom3DtoProtein(aa_3d, residue0, aa_pro, distance0)
}
\arguments{
\item{distance0}{}
}
\description{
this function is used to establish the mapping in the coordinate from protein structure and from protein sequence
}
