% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{mergeActiveSiteInf}
\alias{mergeActiveSiteInf}
\title{Function to summarize all the active site of potein for S. cerevisiae s288c based on
protein function annotation from uniprot database}
\usage{
mergeActiveSiteInf()
}
\description{
Function to summarize all the active site of potein for S. cerevisiae s288c based on
protein function annotation from uniprot database
}
