% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{mergeGeneAnnotationFeature}
\alias{mergeGeneAnnotationFeature}
\title{this function was used to parse the genome annotation of s288c
the result of this script is the base for all other analysis and should be run firstly
firstly we need download gene annotation information from ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/fungi/Saccharomyces_cerevisiae/reference/GCF_000146045.2_R64
which include:
\enumerate{
\item GCF_000146045.2_R64_genomic.gbff
\item GCF_000146045.2_R64_cds_from_genomic.fna
secondly we download the gene annotation information from SGD
\item s288_genome, which include the colnames:  \link{1} "DBID"               "systematic_name"    "gene_standard_name" "gene_name"
"locations_strand"   "locations_start"    "locations_end"      "chromosome"
"sequence"           "sequnece_length"    "protein_residue"    "protein_length"
}}
\usage{
mergeGeneAnnotationFeature()
}
\description{
this function was used to parse the genome annotation of s288c
the result of this script is the base for all other analysis and should be run firstly
firstly we need download gene annotation information from ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/fungi/Saccharomyces_cerevisiae/reference/GCF_000146045.2_R64
which include:
\enumerate{
\item GCF_000146045.2_R64_genomic.gbff
\item GCF_000146045.2_R64_cds_from_genomic.fna
secondly we download the gene annotation information from SGD
\item s288_genome, which include the colnames:  \link{1} "DBID"               "systematic_name"    "gene_standard_name" "gene_name"
"locations_strand"   "locations_start"    "locations_end"      "chromosome"
"sequence"           "sequnece_length"    "protein_residue"    "protein_length"
}
}
