% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{mergeHotspot}
\alias{mergeHotspot}
\title{Title}
\usage{
mergeHotspot(infile, outfile)
}
\arguments{
\item{infile}{the directory contains the hotspot analysis result}

\item{outfile}{the directory contains the hotspot analysis result}
}
\description{
Title
}
