% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_parse_active_site.R
\name{parseSingleSite}
\alias{parseSingleSite}
\title{Function to parse the UniProt annotation with only single site(?)}
\usage{
parseSingleSite(site_dataframe, site_type0)
}
\arguments{
\item{site_dataframe}{a dataframe to store the active site annotation from UniProt database}

\item{site_type0}{the site type classification by UniProt database}
}
\description{
Function to parse the UniProt annotation with only single site(?)
}
