% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_process.R
\name{pdb.ResidueDistance}
\alias{pdb.ResidueDistance}
\title{Title}
\usage{
pdb.ResidueDistance(pdbdir, chainID)
}
\arguments{
\item{pdbdir}{dir for a pdb file}

\item{chainID}{chainID for a chain}
}
\description{
Title
}
