% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_parse.R
\name{pdb.ResolutionEX}
\alias{pdb.ResolutionEX}
\title{This funcion is used to obtain the resolution for the experimetal pdb files
it should be noted, different from experimental PDB files, the resolution of homology pdb files
represents the resolution of related templated PDB files}
\usage{
pdb.ResolutionEX(pdbdir)
}
\arguments{
\item{pdbdir}{directory of a pdb file}
}
\description{
This funcion is used to obtain the resolution for the experimetal pdb files
it should be noted, different from experimental PDB files, the resolution of homology pdb files
represents the resolution of related templated PDB files
}
