% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_parse.R
\name{pdb.Sequence}
\alias{pdb.Sequence}
\title{Get PDB residue sequence}
\usage{
pdb.Sequence(pdbdir, pdbid)
}
\arguments{
\item{pdbdir}{Directory of a pdb file}

\item{pdbid}{Name of the pdb file, like '6cp6.pdb'}
}
\value{
A list contains the residues sequence of each chain in a PDB file
}
\description{
Extract the residue sequence from both experimental and homology pdb files
Please library(bio3d) firstly when using this function
}
\examples{
pdbdir0 <-  "xx/data/6cp6.pdb"
pdb.Sequence(pdbdir=pdbdir0, pdbid='6cp6.pdb')
}
