% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_quality_analysis.R
\name{pdbExFilter}
\alias{pdbExFilter}
\title{experimental pdb file filter function}
\usage{
pdbExFilter(pdb_EX, pident0 = 100, mismatch0 = 0, resolution0 = 3.4)
}
\arguments{
\item{pdb_EX}{a dataframe contains paramters for each experimental pdb file}

\item{pident0}{identity between target protein sequence and the residues sequence included in experimental PDB files}

\item{mismatch0}{mismatch between target protein sequence and the residues sequence included in experimental PDB files}

\item{resolution0}{}
}
\description{
experimental pdb file filter function
}
