% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{plotNullDistribution}
\alias{plotNullDistribution}
\title{Plot the desity graph of WAP}
\usage{
plotNullDistribution(wap_sample, wap_original0 = FALSE)
}
\arguments{
\item{wap_sample}{A vector containing the sampled WAP score}

\item{wap_original0}{A number}
}
\value{
An R graph
}
\description{
Obtain the distribution of sampled WAP
}
