% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{PositionResidue}
\alias{PositionResidue}
\title{part 2 function related to hot spot analysis}
\usage{
PositionResidue(alted_seq, geneName)
}
\arguments{
\item{geneName}{}
}
\description{
this function return the mutated informaiton based on genomics fasta information
this function is used to establish the relation between the mutated amino acid and related position
}
