% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{PositionResidueSNP}
\alias{PositionResidueSNP}
\title{this function return the mutated informaiton based on SNP information: mutated position and changed amino acids}
\usage{
PositionResidueSNP(mutatedPosition, alted, geneName)
}
\arguments{
\item{geneName}{}
}
\description{
this function return the mutated informaiton based on SNP information: mutated position and changed amino acids
}
