% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSNPforGene.R
\name{printSNPforGene}
\alias{printSNPforGene}
\title{function used to give summary of SNP for gene based on a specific strain phenotype}
\usage{
printSNPforGene(gene0 = ss0, SNPlist0 = mutated_gene1,
  gene_annotation0 = gene_feature_GEM, pdbID0 = pdbID, sstart0 = p1,
  send0 = p2)
}
\arguments{
\item{gene0}{A gene systematic name}

\item{SNPlist0}{A SNP list for the strains from specific phenotype}

\item{gene_annotation0}{The gene annotation summary}

\item{pdbID0}{A pdb id which has been mapped onto the gene name (the first parameter)}

\item{sstart0}{The start residue coordinate for the resdiues in the PDB file}

\item{send0}{The end residue coordinate for the residues in the PDB file}
}
\description{
function used to give summary of SNP for gene based on a specific strain phenotype
}
