% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{processFasta}
\alias{processFasta}
\title{this function is used to just preprocess the fasta file without filteration, used for the early version}
\usage{
processFasta(gene_test)
}
\arguments{
\item{gene_test}{}
}
\description{
this function is used to just preprocess the fasta file without filteration, used for the early version
}
