% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{qualityCheckFromCDStoProtein}
\alias{qualityCheckFromCDStoProtein}
\title{this functio was used to evaluate the quality of gene annotation from cds to protein sequence
three check step
check 1, ratio of gene and the translated protein----OK
check 2, length of gene----OK
check 3, translated protein seq is equal to that in SGD, index_need_check 570 1220 1221 1222 1223 1224 1225 1226}
\usage{
qualityCheckFromCDStoProtein(gene_feature_original)
}
\arguments{
\item{gene_feature_original}{}
}
\description{
this functio was used to evaluate the quality of gene annotation from cds to protein sequence
three check step
check 1, ratio of gene and the translated protein----OK
check 2, length of gene----OK
check 3, translated protein seq is equal to that in SGD, index_need_check 570 1220 1221 1222 1223 1224 1225 1226
}
