% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function_in_cluster_analysis.R
\name{removeStopCoden}
\alias{removeStopCoden}
\title{Remove stop coden}
\usage{
removeStopCoden(residue_pair00)
}
\arguments{
\item{residue_pair00}{A dataframe contains the information of sigificant pairs}
}
\value{
A dataframe of sigificant pairs without residue translated by stop coden
}
\description{
Remove the wrong residues obtained by the stop coden
if it is stop coden, then the residue could be *, thus the next step will stop
}
