% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{ResidueSum}
\alias{ResidueSum}
\title{this function is used to put the residue from the same postion together}
\usage{
ResidueSum(pos_residue)
}
\arguments{
\item{pos_residue}{}
}
\description{
this function is used to put the residue from the same postion together
}
