% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{sampleStand}
\alias{sampleStand}
\title{Standardize the mutation number}
\usage{
sampleStand(sample_num)
}
\arguments{
\item{sample_num}{A number}
}
\value{
A number
}
\description{
Calculate the standard samples contained the mutation in the specific postion
}
