% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function_in_cluster_analysis.R
\name{snpFilterBasedMAF}
\alias{snpFilterBasedMAF}
\title{Filter snp value based on maf_value}
\usage{
snpFilterBasedMAF(snp_df, filter_type, maf_value, unique_sign = FALSE)
}
\arguments{
\item{snp_df}{snp_ds is a dataframe obtained based on preprocessSNP}

\item{filter_type}{"smaller" or "bigger"}

\item{maf_value}{MAF of snp}

\item{unique_sign}{"TRUE" or "FALSE". The default is FALSE. If unique_sign is TRUE, the unqiue snp will be returned}
}
\value{

}
\description{
Filter snp value based on maf_value
}
\examples{
data('gene_feature0')
snp_df <- preprocessSNP(gene0 = 'YPR184W', gene_feature = gene_feature0, snp_files_dir0 = "data/gene_snp/")
snp_df1 <- snpFilterBasedMAF(snp_df, filter_type="bigger", maf_value=0.05, unique_sign=FALSE)
}
