% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{annotateSNP}
\alias{annotateSNP}
\title{This a general function used to annotate the small SNP dataset}
\usage{
annotateSNP(snp_input, gene_feature = gene_feature0)
}
\arguments{
\item{snp_input}{a dataframe contains columns of "Chr","Pos","Gene","Ref","Alt" for each SNP}

\item{gene_feature}{a dataframe contains the detailed information of a gene}
}
\description{
This a general function used to annotate the small SNP dataset
}
