% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{checkTanslatedProtein}
\alias{checkTanslatedProtein}
\title{this function is used to check whether the translated protein is equal to the protein sequence
from the sgd database, if they are equal, the original cds sequence and protein sequence is consistent in
our program}
\usage{
checkTanslatedProtein(gene_seq_inf)
}
\arguments{
\item{gene_seq_inf}{}
}
\description{
this function is used to check whether the translated protein is equal to the protein sequence
from the sgd database, if they are equal, the original cds sequence and protein sequence is consistent in
our program
}
