% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{chooseStrain}
\alias{chooseStrain}
\title{newly added function, will integrate the main function
this function is mainly used to choose the strain based on the strain type defined in 1011 genome sequence project
example
strain_type <- "all_strain"
strain_select1 <- chooseStrain(type = strain_type)}
\usage{
chooseStrain(type, strain0 = strain_classification)
}
\arguments{
\item{strain0}{}
}
\description{
newly added function, will integrate the main function
this function is mainly used to choose the strain based on the strain type defined in 1011 genome sequence project
example
strain_type <- "all_strain"
strain_select1 <- chooseStrain(type = strain_type)
}
