% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUMPS.R
\name{clumpsAnalysis}
\alias{clumpsAnalysis}
\title{CLUMPS function used to run the CLUMPS pipeline}
\usage{
clumpsAnalysis(gene0, SNPlist0, gene_annotation0, pdb, sstart0, send0,
  input_dir = TRUE)
}
\arguments{
\item{gene0}{A gene systematic name}

\item{SNPlist0}{A SNP list for the strains from specific phenotype}

\item{gene_annotation0}{The gene annotation summary}

\item{pdb}{The dir store the residue distance matrix, should be a txt file seperated by ',' or it is a residue distance matrix}

\item{sstart0}{The start residue coordinate for the resdiues in the PDB file}

\item{send0}{The end residue coordinate for the residues in the PDB file}

\item{input_dir}{}
}
\description{
CLUMPS function used to run the CLUMPS pipeline
}
