% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{Cluster summary}
\usage{
clusterAnalysis(residueInf)
}
\arguments{
\item{residueInf}{A dataframe contains the detailed information about the residue}
}
\value{
A dataframe contains the cluster information
}
\description{
This function is used to obtain the cluster analysis results
}
