% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{this function is used to obtain the cluster analysis results}
\usage{
clusterAnalysis(residueInf)
}
\arguments{
\item{residueInf}{}
}
\description{
this function is used to obtain the cluster analysis results
}
