% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{countMutationProtein}
\alias{countMutationProtein}
\title{the followed two function  were used to estimate the mutation information based on single SNP information
using function to obtain the each gene's mutation information based on the processed mutation data
----------------vesion 2}
\usage{
countMutationProtein(gene_name, mutation_annotation, gene_snp0)
}
\arguments{
\item{mutation_annotation}{}
}
\description{
the followed two function  were used to estimate the mutation information based on single SNP information
using function to obtain the each gene's mutation information based on the processed mutation data
----------------vesion 2
}
