% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getGeneNameWithSNP}
\alias{getGeneNameWithSNP}
\title{List genes with SNPs}
\usage{
getGeneNameWithSNP()
}
\value{
A vector contain the gene name
}
\description{
Obtain the gene list which have SNP, in total there are 36 metabolic genes which
don't have the mutation
}
\examples{
# Firstly, open one R project
# Then put the SNP file in the directory of "xx/data"
getGeneNameWithSNP()
}
