% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getHotPvalue}
\alias{getHotPvalue}
\title{this function is used to calculate p value for each clust}
\usage{
getHotPvalue(cluster0, sample_standard, distance, seq)
}
\arguments{
\item{seq}{}
}
\description{
this function is used to calculate p value for each clust
}
