% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function_in_cluster_analysis.R
\name{getHotVertice}
\alias{getHotVertice}
\title{Get significant pairs}
\usage{
getHotVertice(aa_3d, residue0, aa_pro, distance0)
}
\arguments{
\item{aa_3d}{A vector for the coordinate of PDB structure}

\item{residue0}{A vector contained all the muated residue information of the stucture and it can be found the same mutation in residue occured many times}

\item{aa_pro}{A vector for the original coordinate of protein aa sequence}

\item{distance0}{A matrix for the distance of the paired residue of pdb structure}
}
\value{
A dataframe contains the inforation of the mutated residues
}
\description{
This function is used to filter significant pairs based on hotspot paper
}
