% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getProteinFromCDS}
\alias{getProteinFromCDS}
\title{Translate CDS into Protein}
\usage{
getProteinFromCDS(cds0)
}
\arguments{
\item{cds0}{A string containing cds sequence}
}
\value{
A string containing the protein sequence
}
\description{
Produce the amino acid seq based on cds sequence
}
\examples{
getProteinFromCDS(cds0="ATGGAAATGA")
}
