% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getPvalue}
\alias{getPvalue}
\title{Calculate p-value}
\usage{
getPvalue(wap_initial, wap_sampling)
}
\arguments{
\item{wap_initial}{A number}

\item{wap_sampling}{A vector of wap obtained sampling method}
}
\value{
right tailed p value
}
\description{
Calculate the right-tailed p value
}
