% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{getSampleWAP}
\alias{getSampleWAP}
\title{function to calculate the sample WAP}
\usage{
getSampleWAP(mutated_pos, sample0, distance, seq = seq0, n = 10000)
}
\arguments{
\item{n}{}
}
\description{
function to calculate the sample WAP
}
