% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{getSingleMatchParameter}
\alias{getSingleMatchParameter}
\title{Get single mapping result for one element}
\usage{
getSingleMatchParameter(description, reaction, ko)
}
\arguments{
\item{description}{A vector contains the description for each element from 'reaction'}

\item{reaction}{A vector contrains the element been mapped}

\item{ko}{A vector contains the element need mapping}
}
\value{
A vector
}
\description{
Get single mapping result for one element
}
\examples{
w <- 1:6
v <- c('z','a','b','a','b', 'e')
testData <- c('a','b','g')
getSingleMatchParameter(w,v,testData)

}
