% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{getTotalWAP}
\alias{getTotalWAP}
\title{function to calculate the initial WAP value}
\usage{
getTotalWAP(mutated_pos, sample0, distance)
}
\arguments{
\item{distance}{}
}
\description{
function to calculate the initial WAP value
}
