% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_analysis.R
\name{hotSpotAnalysis}
\alias{hotSpotAnalysis}
\title{Mutation hotspot analysis}
\usage{
hotSpotAnalysis(gene0 = ss0, pdbID = NA, SNPlist0 = mutated_gene1,
  gene_annotation0 = gene_feature0, pdb = distance, sstart0 = p1,
  send0 = p2, qstart0 = q1, qend0 = q2, result_dir = outfile0,
  strain_type = NA, input_dir = TRUE)
}
\arguments{
\item{gene0}{A gene systematic name}

\item{pdbID}{A string}

\item{SNPlist0}{A SNP list for the strains from specific phenotype}

\item{gene_annotation0}{The gene annotation summary}

\item{pdb}{The dir store the residue distance matrix, should be a txt file seperated by ',' or the residue distance matrix}

\item{sstart0}{The start residue coordinate for the resdiues in the original protein}

\item{send0}{The end residue coordinate for the residues in the the original protein}

\item{qstart0}{The start residue coordinate for the resdiues in the PDB file}

\item{qend0}{The end residue coordinate for the residues in the PDB file}

\item{result_dir}{The directory to save the hot spot analysis result}

\item{strain_type}{A string to represent the source of SNP}

\item{input_dir}{A logical vector}
}
\value{
A dataframe contains the hotspot analysis result
}
\description{
Conduct the mutation hotspot analysis based on protein 3D structures
}
\examples{
# Load SNP data and residue distance matrix
data("snp_YBR046C")
data("ResidueDistance_YBR046C")
# Creat file to store the result
outfile0 <- "result/hot_spot_analysis"
dir.create(outfile0)
hotSpotAnalysis(
gene0 = "YBR046C",
 SNPlist0 = snp_YBR046C,
 gene_annotation0 = gene_feature0,
 pdb = ResidueDistance_YBR046C,
 sstart0 = 5, # coordinate of original protein residues sequence
 send0 = 333, # coordinate of original protein residues sequence
 qstart0 = 1, # coordinate of protein residues sequence in pdb file
 qend0 = 329, # coordinate of protein residues sequence in pdb file
 result_dir = outfile0,
 input_dir = FALSE
)
}
