% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package.R
\name{loadPackageForSNPanalysis}
\alias{loadPackageForSNPanalysis}
\title{Load essential packages}
\usage{
loadPackageForSNPanalysis()
}
\value{
Display whether all essential packages are installed.
}
\description{
Load the package used for SNP analysis
}
\examples{
library(Yeastspot3D)
loadPackageForSNPanalysis()
}
