% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{mergeHotspot}
\alias{mergeHotspot}
\title{Merge hotspots}
\usage{
mergeHotspot(infile, outfile)
}
\arguments{
\item{infile}{A directory contains the hotspot analysis result}

\item{outfile}{A directory contains the hotspot analysis result}
}
\value{
A txt file contains all hotspots
}
\description{
Put all the hotspot into txt file
}
\examples{
# choose the directory contains the hotspot analysis result
mergeHotspot (infile = 'result/hotspot analysis'
              outfile = 'result/hotspot analysis')

}
