% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_parse_active_site.R
\name{parseMutipleSite}
\alias{parseMutipleSite}
\title{Function to parse the UniProt annotation with only single or several site(?)}
\usage{
parseMutipleSite(site_dataframe, site_type0, single_site = TRUE)
}
\arguments{
\item{single_site}{}
}
\description{
Function to parse the UniProt annotation with only single or several site(?)
}
