% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_parse.R
\name{pdb.ResidueDistance}
\alias{pdb.ResidueDistance}
\title{Calculate the distance between each two alpha carbons}
\usage{
pdb.ResidueDistance(pdbdir, chainID)
}
\arguments{
\item{pdbdir}{Dir for a pdb file}

\item{chainID}{ChainID for a chain}
}
\value{
A matrix contains the distance between each two alpha carbons
}
\description{
Calculate the distance between each two alpha carbons
}
\examples{
library(bio3d)
library(seqinr)
infile <- "xx/data/"
pdbid <- '6cp6.pdb'
pdbdir <- paste(infile, pdbid, sep = "")
pdb.ResidueDistance(pdbdir, chainID = 'K')
}
