% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_quality_analysis.R
\name{pdb.ResolutionAll}
\alias{pdb.ResolutionAll}
\title{Function to parse the pdb web https://www.rcsb.org/ to obtain the resolution for each pdb file
Please firstly library(rvest) when using this function}
\usage{
pdb.ResolutionAll(pdb0)
}
\arguments{
\item{pdb0}{a pdb id}
}
\description{
Function to parse the pdb web https://www.rcsb.org/ to obtain the resolution for each pdb file
Please firstly library(rvest) when using this function
}
