% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_parse.R
\name{pdb.ResolutionEX}
\alias{pdb.ResolutionEX}
\title{Obtain the resolution for one experimetal PDB file}
\usage{
pdb.ResolutionEX(pdbdir)
}
\arguments{
\item{pdbdir}{Directory of a pdb file}
}
\value{
Resolution for the pdb file
}
\description{
Get the resolution of an experimental PDB file
}
\examples{
pdbdir0 <- "xx/data/6cp6.pdb"
pdb.ResolutionEX(pdbdir=pdbdir0)
}
