% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_process.R
\name{pdb.Sequence}
\alias{pdb.Sequence}
\title{This function is used to extract the residue sequence from both experimental and homology pdb files
Please library(bio3d) firstly when using this function}
\usage{
pdb.Sequence(pdbdir, pdbid)
}
\arguments{
\item{pdbdir}{directory of a pdb file}

\item{pdbid}{name of the pdb file, like '6cp6.pdb'}
}
\description{
This function is used to extract the residue sequence from both experimental and homology pdb files
Please library(bio3d) firstly when using this function
}
