% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_quality_analysis.R
\name{pdbHomoFilter}
\alias{pdbHomoFilter}
\title{homology pdb file filter function}
\usage{
pdbHomoFilter(pdb_homo, qmean0 = -4, identity0 = 25,
  similarity0 = 0.31, resolution0 = 3.4)
}
\arguments{
\item{pdb_homo}{a dataframe contains paramters for each homology pdb file}

\item{qmean0}{paramter to evaluate the pdb quality defined in Swiss model database}

\item{identity0}{Seq-identity, identity between target protein sequence and the residues sequence of templated PDB files}

\item{similarity0}{Seq_similarity, similarity between target protein sequence and the residues sequence of templated PDB files}

\item{resolution0}{Resolution}
}
\description{
homology pdb file filter function
}
