% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{plotNullDistribution}
\alias{plotNullDistribution}
\title{function to plot the desity graph
the wap_original0 now can be added as a line in x-axis}
\usage{
plotNullDistribution(wap_sample, wap_original0 = FALSE)
}
\arguments{
\item{wap_original0}{}
}
\description{
function to plot the desity graph
the wap_original0 now can be added as a line in x-axis
}
