% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{preprocessSNP}
\alias{preprocessSNP}
\title{this function is used to proprocess all the snp for one gene belong to a sample set
it should be noted that if the gene belong to minus strand, changeATCG function will be used
be careful about the file directory}
\usage{
preprocessSNP(gene0)
}
\arguments{
\item{gene0}{}
}
\description{
this function is used to proprocess all the snp for one gene belong to a sample set
it should be noted that if the gene belong to minus strand, changeATCG function will be used
be careful about the file directory
}
