% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{printSNPforGene}
\alias{printSNPforGene}
\title{Print the mutated residues}
\usage{
printSNPforGene(
  gene0 = ss0,
  SNPlist0 = mutated_gene1,
  gene_annotation0 = gene_feature0,
  pdbID0 = pdbID,
  sstart0 = p1,
  send0 = p2
)
}
\arguments{
\item{gene0}{A gene systematic name}

\item{SNPlist0}{A SNP list for the strains from specific phenotype}

\item{gene_annotation0}{The gene annotation summary}

\item{pdbID0}{A pdb id which has been mapped onto the gene name (the first parameter)}

\item{sstart0}{The start residue coordinate for the resdiues in the PDB file}

\item{send0}{The end residue coordinate for the residues in the PDB file}
}
\value{
A dataframe contains the mutated residues information
}
\description{
Output all the mutated residues information
}
\examples{
data('gene_feature0')
data('snp_data')
mutated_gene <- annotateSNP(snp_input = snp_data, gene_feature = gene_feature0)
data('ResidueDistance_YPR184W')
mutated_gene1 <- filter(mutated_gene, Gene2 == 'YPR184W')
pdbID <- '2_1534_5d06.1.A_5b2453487f4bf94bf75ead43'
SNP_list <- printSNPforGene(gene0 = 'YPR184W',
                            SNPlist0 = mutated_gene1,
                            gene_annotation0 = gene_feature0,
                            pdbID0 = pdbID,
                            sstart0 = 2,
                            send0 = 1534)
}
