% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{removeStopCoden}
\alias{removeStopCoden}
\title{this function is used to remove the wrong residues obtained by the stop coden
if it is stop coden, then the residue could be *, thus the next step will stop}
\usage{
removeStopCoden(residue_pair00)
}
\arguments{
\item{residue_pair00}{}
}
\description{
this function is used to remove the wrong residues obtained by the stop coden
if it is stop coden, then the residue could be *, thus the next step will stop
}
