% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function_in_cluster_analysis.R
\name{ResidueSum}
\alias{ResidueSum}
\title{Summarize the mutated residues information}
\usage{
ResidueSum(pos_residue)
}
\arguments{
\item{pos_residue}{A list contains the mutated residues information}
}
\value{
A dataframe contains the mutation position and the 'alt' residues
}
\description{
Put the residue from the same postion together
}
