% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function_test.R
\name{sampleStand}
\alias{sampleStand}
\title{fuction to calculate the standard samples contained the mutation in the specific postion}
\usage{
sampleStand(sample_num)
}
\arguments{
\item{sample_num}{}
}
\description{
fuction to calculate the standard samples contained the mutation in the specific postion
}
