% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R, R/main_function.R
\name{splitAndCombine}
\alias{splitAndCombine}
\title{Fast estabolish the mapping relation between two element}
\usage{
splitAndCombine(gene, rxn, sep0)

splitAndCombine(gene, rxn, sep0)
}
\arguments{
\item{gene}{A vetor contains elements with "&" or ","}

\item{rxn}{A vector contains unique identifier}

\item{sep0}{A string, i.e, "&". ","}

\item{sep0}{}
}
\value{
A dataframe contains the single mapping relation for the element from gene and rxn
}
\description{
Fast estabolish the mapping relation between two element

this function is from hongR
}
\examples{
gene <- c('a&b','c')
rxn <- c('r1','r2')
splitAndCombine(gene, rxn, sep0="&")

}
